change all ids to long to match the dbs column type

+++ src/main/java/ch/bfh/bti7081/s2017/grey/database/dao/AppointmentDao.java
@@ -68,7 +68,7 @@ public class AppointmentDao {
         emfactory.close();
     }
 
-    public void removeAppointment(int id) {
+    public void removeAppointment(long id) {
         EntityManagerFactory emfactory = Persistence.createEntityManagerFactory("CRM");
         EntityManager entitymanager = emfactory.createEntityManager();
         entitymanager.getTransaction().begin();
