+++ src/main/java/ch/bfh/bti7081/s2017/grey/service/PatientService.java
@@ -6,7 +6,7 @@ import ch.bfh.bti7081.s2017.grey.database.entity.Patient;
  * @Author Quentin
  */
 public interface PatientService {
-    Patient getPatientById(int id);
+    Patient getPatientById(long id);
     void createStaff(String firstname, String lastname);
     Patient getPatientByName(String firstName, String lastName);
 }
