+++ src/main/java/ch/bfh/bti7081/s2017/grey/service/impl/PatientServiceImpl.java
@@ -15,7 +15,7 @@ public class PatientServiceImpl implements PatientService {
         patientDao = new PatientDao();
     }
     @Override
-    public Patient getPatientById(int id) {
+    public Patient getPatientById(long id) {
         return patientDao.getPatientById(id);
     }
